LACE Processing and Analysis Interface

LACEview displays a Shiny user interface to handle the VCF and BAM files processing that is needed to construct the input for the LACE inference algorithms. The function generates also the maximum likelihood longitudinal clonal tree, and shows the output for further explorations of the results.

LACEview()

Value

The GUI

Note

The function LACE is still available for retrocompatibility.

Installation

The package is available on GitHub and Bioconductor. LACE 2.0 requires R > 4.1.0 and Bioconductor.

To install directly from github run:

remotes::install_github("https://github.com/BIMIB-DISCo/LACE", 
                        dependencies = TRUE)

Dependencies

LACE 2.0 uses Annovar and Samtools suite as back-ends for variant calling annotation and depth computation, respectively.

Annovar is a variant calling software written in Perl freely available upon registration to their website at https://annovar.openbioinformatics.org/en/latest/.

Perl can be found and installed at https://www.perl.org/.

Samtools suite is a set of tools to handle SAM/BAM/BED file format. It is freely available at http://www.htslib.org/. To install Samtools follow the instructions in their website.