lacedata

Perform the inference of the maximum likelihood clonal tree from longitudinal data.

lacedata(
  D,
  lik_w = NULL,
  alpha = NULL,
  beta = NULL,
  initialization = NULL,
  random_tree = FALSE,
  keep_equivalent = TRUE,
  check_indistinguishable = TRUE,
  num_rs = 50,
  num_iter = 10000,
  n_try_bs = 500,
  learning_rate = 1,
  marginalize = FALSE,
  error_move = FALSE,
  num_processes = Inf,
  seed = NULL,
  verbose = TRUE,
  log_file = ""
)

Arguments

D

Mutation data from multiple experiments for a list of driver genes. It can be either a list with a data matrix per time point or a SummarizedExperiment object. In this latter, the object must contain two fields: assays and colData. Assays stores one unique data matrix pooling all single cells observed at each time point and colData stores a vector of labels reporting the time point when each single cell was sequenced. Ordering of cells in assays field and colData field must be the same.

lik_w

Weight for each data point. If not provided, weights to correct for sample sizes are used.

alpha

False positive error rate provided as list of elements; if a vector of alpha (and beta) is provided, the inference is performed for multiple values and the solution at maximum-likelihood is returned.

beta

False negative error rate provided as list of elements; if a vector of beta (and alpha) is provided, the inference is performed for multiple values and the solution at maximum-likelihood is returned.

initialization

Binary matrix representing a perfect philogeny clonal tree; clones are rows and mutations are columns. This parameter overrides "random_tree".

random_tree

Boolean. Shall I start MCMC search from a random tree? If FALSE (default) and initialization is NULL, search is started from a TRaIT tree (BMC Bioinformatics . 2019 Apr 25;20(1):210. doi: 10.1186/s12859-019-2795-4).

keep_equivalent

Boolean. Shall I return results (B and C) at equivalent likelihood with the best returned solution?

check_indistinguishable

Boolean. Shall I remove any indistinguishable event from input data prior inference?

num_rs

Number of restarts during mcmc inference.

num_iter

Maximum number of mcmc steps to be performed during the inference.

n_try_bs

Number of steps without change in likelihood of best solution after which to stop the mcmc.

learning_rate

Parameter to tune the probability of accepting solutions at lower values during mcmc. Value of learning_rate = 1 (default), set a probability proportional to the difference in likelihood; values of learning_rate greater than 1 inclease the chance of accepting solutions at lower likelihood during mcmc while values lower than 1 decrease such probability.

marginalize

Boolean. Shall I marginalize C when computing likelihood?

error_move

Boolean. Shall I include estimation of error rates in the MCMC moves?

num_processes

Number of processes to be used during parallel execution. To execute in single process mode, this parameter needs to be set to either NA or NULL.

seed

Seed for reproducibility.

verbose

Boolean. Shall I print to screen information messages during the execution?

log_file

log file where to print outputs when using parallel. If parallel execution is disabled, this parameter is ignored.

Value

shiny interface

Examples

data(longitudinal_sc_variants)
lacedata(D = longitudinal_sc_variants,
         lik_w = c(0.2308772,0.2554386,0.2701754,0.2435088),
         alpha = list(c(0.10,0.05,0.05,0.05)),
         beta = list(c(0.10,0.05,0.05,0.05)),
         keep_equivalent = TRUE,
         num_rs = 5,
         num_iter = 10,
         n_try_bs = 5,
         num_processes = NA,
         seed = 12345,
         verbose = FALSE)
#> [1] "B is already binary"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "nodes in B are sprouting from a single node"
#> Log is not open.
#> [1] "B is a full-rank matrix."
#> Log is not open.
#> [1] "Root Root 0"
#> Log is not open.
#> [1] "Root Clone_1 1"
#> Log is not open.
#> [1] "Root Clone_2 0"
#> Log is not open.
#> [1] "Root Clone_3 0"
#> Log is not open.
#> [1] "Root Clone_4 0"
#> Log is not open.
#> [1] "Root Clone_5 0"
#> Log is not open.
#> [1] "Root Clone_6 0"
#> Log is not open.
#> [1] "Clone_1 Clone_1 1"
#> Log is not open.
#> [1] "Clone_1 Clone_2 1"
#> Log is not open.
#> [1] "Clone_1 Clone_3 0"
#> Log is not open.
#> [1] "Clone_1 Clone_4 0"
#> Log is not open.
#> [1] "Clone_1 Clone_5 0"
#> Log is not open.
#> [1] "Clone_1 Clone_6 0"
#> Log is not open.
#> [1] "Clone_2 Clone_2 1"
#> Log is not open.
#> [1] "Clone_2 Clone_3 1"
#> Log is not open.
#> [1] "Clone_2 Clone_4 1"
#> Log is not open.
#> [1] "Clone_2 Clone_5 0"
#> Log is not open.
#> [1] "Clone_2 Clone_6 0"
#> Log is not open.
#> [1] "Clone_3 Clone_3 1"
#> Log is not open.
#> [1] "Clone_3 Clone_4 1"
#> Log is not open.
#> [1] "Clone_3 Clone_5 1"
#> Log is not open.
#> [1] "Clone_3 Clone_6 0"
#> Log is not open.
#> [1] "Clone_4 Clone_4 1"
#> Log is not open.
#> [1] "Clone_4 Clone_5 1"
#> Log is not open.
#> [1] "Clone_4 Clone_6 0"
#> Log is not open.
#> [1] "Clone_5 Clone_5 1"
#> Log is not open.
#> [1] "Clone_5 Clone_6 1"
#> Log is not open.
#> [1] "Clone_6 Clone_6 1"
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#> [1] TRUE
#> Log is not open.
#> [1] "B represents a forest"
#> Log is not open.
#> [1] "B has a forest like structure"
#> Log is not open.
#> [1] "Root Root 0"
#> Log is not open.
#> [1] "Root Clone_1 1"
#> Log is not open.
#> [1] "Root Clone_2 0"
#> Log is not open.
#> [1] "Root Clone_3 0"
#> Log is not open.
#> [1] "Root Clone_4 0"
#> Log is not open.
#> [1] "Root Clone_5 0"
#> Log is not open.
#> [1] "Root Clone_6 0"
#> Log is not open.
#> [1] "Clone_1 Clone_1 1"
#> Log is not open.
#> [1] "Clone_1 Clone_2 1"
#> Log is not open.
#> [1] "Clone_1 Clone_3 0"
#> Log is not open.
#> [1] "Clone_1 Clone_4 0"
#> Log is not open.
#> [1] "Clone_1 Clone_5 0"
#> Log is not open.
#> [1] "Clone_1 Clone_6 0"
#> Log is not open.
#> [1] "Clone_2 Clone_2 1"
#> Log is not open.
#> [1] "Clone_2 Clone_3 1"
#> Log is not open.
#> [1] "Clone_2 Clone_4 1"
#> Log is not open.
#> [1] "Clone_2 Clone_5 0"
#> Log is not open.
#> [1] "Clone_2 Clone_6 0"
#> Log is not open.
#> [1] "Clone_3 Clone_3 1"
#> Log is not open.
#> [1] "Clone_3 Clone_4 1"
#> Log is not open.
#> [1] "Clone_3 Clone_5 1"
#> Log is not open.
#> [1] "Clone_3 Clone_6 0"
#> Log is not open.
#> [1] "Clone_4 Clone_4 1"
#> Log is not open.
#> [1] "Clone_4 Clone_5 1"
#> Log is not open.
#> [1] "Clone_4 Clone_6 0"
#> Log is not open.
#> [1] "Clone_5 Clone_5 1"
#> Log is not open.
#> [1] "Clone_5 Clone_6 1"
#> Log is not open.
#> [1] "Clone_6 Clone_6 1"
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#> [1] TRUE
#> Log is not open.
#> [1] "B represents a forest"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "B is a binary matrix"
#> Log is not open.
#> [1] "B is a square matrix"
#> Log is not open.
#> [1] "B is a full rank matrix"
#> Log is not open.
#> [1] "B is a forest "
#> Log is not open.
#> [1] "B has single root"
#> Log is not open.
#> [1] "all checks done!"
#> Log is not open.
#> [1] "continue..."
#> Log is not open.
#> [1] "clonal tree:"
#> Log is not open.
#>                        Root ARPC2_2_218249894_C_T PRAME_22_22551005_T_A
#> Root                      0                     1                     0
#> ARPC2_2_218249894_C_T     0                     0                     1
#> PRAME_22_22551005_T_A     0                     0                     0
#> COL1A2_7_94422978_C_A     0                     0                     0
#> HNRNPC_14_21211843_C_T    0                     0                     0
#> CCT8_21_29063389_G_A      0                     0                     0
#> RPL5_1_92837514_C_G       0                     0                     0
#>                        COL1A2_7_94422978_C_A HNRNPC_14_21211843_C_T
#> Root                                       0                      0
#> ARPC2_2_218249894_C_T                      0                      0
#> PRAME_22_22551005_T_A                      1                      1
#> COL1A2_7_94422978_C_A                      0                      0
#> HNRNPC_14_21211843_C_T                     0                      0
#> CCT8_21_29063389_G_A                       0                      0
#> RPL5_1_92837514_C_G                        0                      0
#>                        CCT8_21_29063389_G_A RPL5_1_92837514_C_G
#> Root                                      0                   0
#> ARPC2_2_218249894_C_T                     0                   0
#> PRAME_22_22551005_T_A                     0                   0
#> COL1A2_7_94422978_C_A                     1                   0
#> HNRNPC_14_21211843_C_T                    0                   0
#> CCT8_21_29063389_G_A                      0                   1
#> RPL5_1_92837514_C_G                       0                   0
#> Log is not open.
#>         Root ARPC2_2_218249894_C_T PRAME_22_22551005_T_A COL1A2_7_94422978_C_A
#> Root       1                     0                     0                     0
#> Clone_1    1                     1                     0                     0
#> Clone_2    1                     1                     1                     0
#> Clone_3    1                     1                     1                     1
#> Clone_4    1                     1                     1                     0
#> Clone_5    1                     1                     1                     1
#> Clone_6    1                     1                     1                     1
#>         HNRNPC_14_21211843_C_T CCT8_21_29063389_G_A RPL5_1_92837514_C_G
#> Root                         0                    0                   0
#> Clone_1                      0                    0                   0
#> Clone_2                      0                    0                   0
#> Clone_3                      0                    0                   0
#> Clone_4                      1                    0                   0
#> Clone_5                      0                    1                   0
#> Clone_6                      0                    1                   1
#> Log is not open.
#> [1] "No incongruences found due to the chronological order"
#> Log is not open.
#> [1] "first time mutation occurrences based the chronological order of samples:"
#> Log is not open.
#>    Root Clone_1 Clone_2 Clone_3 Clone_4 Clone_5 Clone_6 
#>       1       1       1       1       1       1       1 
#> Log is not open.
#> [1] "adjacent matrix with first time occurrences:"
#> Log is not open.
#>            Root_t1 Clone_1_t1 Clone_2_t1 Clone_3_t1 Clone_4_t1 Clone_5_t1
#> Root_t1          0          1          0          0          0          0
#> Clone_1_t1       0          0          1          0          0          0
#> Clone_2_t1       0          0          0          1          1          0
#> Clone_3_t1       0          0          0          0          0          1
#> Clone_4_t1       0          0          0          0          0          0
#> Clone_5_t1       0          0          0          0          0          0
#> Clone_6_t1       0          0          0          0          0          0
#>            Clone_6_t1
#> Root_t1             0
#> Clone_1_t1          0
#> Clone_2_t1          0
#> Clone_3_t1          0
#> Clone_4_t1          0
#> Clone_5_t1          1
#> Clone_6_t1          0
#> Log is not open.
#> [1] "time=" "1"    
#> Log is not open.
#> [1] "1st Root_t1->Root_t1=1, prev_Root_t1=0.143835616438356: Root_t1->Root_t2=2"
#> Log is not open.
#> [1] "1st Root_t1->Root_t2=2, prev_Root_t2=0.162162162162162: Root_t2->Root_t3=2"
#> Log is not open.
#> [1] "1st Root_t1->Clone_1_t1=1, prev_Clone_1_t1=0.157534246575342: Clone_1_t1->Clone_1_t2=2"
#> Log is not open.
#> [1] "1st Clone_1_t1->Clone_1_t2=2, prev_Clone_1_t2=0.225225225225225: Clone_1_t2->Clone_1_t3=2"
#> Log is not open.
#> [1] "1st Clone_1_t1->Clone_2_t1=1, prev_Clone_2_t1=0.184931506849315: Clone_2_t1->Clone_2_t2=2"
#> Log is not open.
#> [1] "1st Clone_2_t1->Clone_2_t2=2, prev_Clone_2_t2=0.243243243243243: Clone_2_t2->Clone_2_t3=2"
#> Log is not open.
#> [1] "1st Clone_2_t1->Clone_3_t1=1, prev_Clone_3_t1=0.280821917808219: Clone_3_t1->Clone_3_t2=2"
#> Log is not open.
#> [1] "1st Clone_2_t1->Clone_4_t1=1, prev_Clone_4_t1=0.0479452054794521: Clone_4_t1->Clone_4_t2=2"
#> Log is not open.
#> [1] "1st Clone_3_t1->Clone_3_t2=2, prev_Clone_3_t2=0.162162162162162: Clone_3_t2->Clone_3_t3=2"
#> Log is not open.
#> [1] "1st Clone_3_t1->Clone_5_t1=1, prev_Clone_5_t1=0.0958904109589041: Clone_5_t1->Clone_5_t2=2"
#> Log is not open.
#> [1] "1st Clone_4_t1->Clone_4_t2=2, prev_Clone_4_t2=0.0810810810810811: Clone_4_t2->Clone_4_t3=2"
#> Log is not open.
#> [1] "1st Clone_5_t1->Clone_5_t2=2, prev_Clone_5_t2=0.126126126126126: Clone_5_t2->Clone_5_t3=2"
#> Log is not open.
#> [1] "1st Clone_5_t1->Clone_6_t1=1, prev_Clone_6_t1=0.089041095890411: Clone_6_t1->Clone_6_t2=2"
#> Log is not open.
#> [1] "Clone_6_t1->Clone_6_t2=2, prev_Clone_6_t2=0: Clone_6_t1->Clone_6_t2=3"
#> Log is not open.
#> [1] "Clone_6_t1->Clone_6_t2=3, prev_Clone_6_t2=0: Clone_6_t2->Clone_6_t3=3"
#> Log is not open.
#> [1] "time=" "2"    
#> Log is not open.
#> [1] "1st Root_t2->Root_t3=2, prev_Root_t3=0.133333333333333: Root_t3->Root_t4=2"
#> Log is not open.
#> [1] "1st Clone_1_t2->Clone_1_t3=2, prev_Clone_1_t3=0.166666666666667: Clone_1_t3->Clone_1_t4=2"
#> Log is not open.
#> [1] "1st Clone_2_t2->Clone_2_t3=2, prev_Clone_2_t3=0.166666666666667: Clone_2_t3->Clone_2_t4=2"
#> Log is not open.
#> [1] "1st Clone_3_t2->Clone_3_t3=2, prev_Clone_3_t3=0.255555555555556: Clone_3_t3->Clone_3_t4=2"
#> Log is not open.
#> [1] "1st Clone_4_t2->Clone_4_t3=2, prev_Clone_4_t3=0.0333333333333333: Clone_4_t3->Clone_4_t4=2"
#> Log is not open.
#> [1] "1st Clone_5_t2->Clone_5_t3=2, prev_Clone_5_t3=0.211111111111111: Clone_5_t3->Clone_5_t4=2"
#> Log is not open.
#> [1] "1st Clone_6_t2->Clone_6_t3=3, prev_Clone_6_t3=0.0333333333333333: Clone_6_t3->Clone_6_t4=2"
#> Log is not open.
#> [1] "time=" "3"    
#> Log is not open.
#> [1] "time=" "4"    
#> Log is not open.
#> [1] "time=" "3"    
#> Log is not open.
#> [1] "time=" "2"    
#> Log is not open.
#> [1] "time=" "1"    
#> Log is not open.
#> [1] "idx"     "Clone_1"
#> Log is not open.
#> [1] "idx"     "Clone_2"
#> Log is not open.
#> [1] "idx"     "Clone_3"
#> Log is not open.
#> [1] "idx"     "Clone_4"
#> Log is not open.
#> [1] "idx"     "Clone_5"
#> Log is not open.
#> [1] "idx"     "Clone_6"
#> Log is not open.
#> [1] "/tmp/RtmpLExdXL/file5df03d88cab1/reference"
#> Log is not open.
#> [1] "info"
#> Log is not open.
#> [1] ""
#> Log is not open.
#> [1] "B is already binary"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "nodes in B are sprouting from a single node"
#> Log is not open.
#> [1] "B is a full-rank matrix."
#> Log is not open.
#> [1] "Root Root 0"
#> Log is not open.
#> [1] "Root Clone_1 1"
#> Log is not open.
#> [1] "Root Clone_2 0"
#> Log is not open.
#> [1] "Root Clone_3 0"
#> Log is not open.
#> [1] "Root Clone_4 0"
#> Log is not open.
#> [1] "Root Clone_5 0"
#> Log is not open.
#> [1] "Root Clone_6 0"
#> Log is not open.
#> [1] "Clone_1 Clone_1 1"
#> Log is not open.
#> [1] "Clone_1 Clone_2 1"
#> Log is not open.
#> [1] "Clone_1 Clone_3 0"
#> Log is not open.
#> [1] "Clone_1 Clone_4 0"
#> Log is not open.
#> [1] "Clone_1 Clone_5 0"
#> Log is not open.
#> [1] "Clone_1 Clone_6 0"
#> Log is not open.
#> [1] "Clone_2 Clone_2 1"
#> Log is not open.
#> [1] "Clone_2 Clone_3 1"
#> Log is not open.
#> [1] "Clone_2 Clone_4 1"
#> Log is not open.
#> [1] "Clone_2 Clone_5 0"
#> Log is not open.
#> [1] "Clone_2 Clone_6 0"
#> Log is not open.
#> [1] "Clone_3 Clone_3 1"
#> Log is not open.
#> [1] "Clone_3 Clone_4 1"
#> Log is not open.
#> [1] "Clone_3 Clone_5 1"
#> Log is not open.
#> [1] "Clone_3 Clone_6 0"
#> Log is not open.
#> [1] "Clone_4 Clone_4 1"
#> Log is not open.
#> [1] "Clone_4 Clone_5 1"
#> Log is not open.
#> [1] "Clone_4 Clone_6 0"
#> Log is not open.
#> [1] "Clone_5 Clone_5 1"
#> Log is not open.
#> [1] "Clone_5 Clone_6 1"
#> Log is not open.
#> [1] "Clone_6 Clone_6 1"
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#> [1] TRUE
#> Log is not open.
#> [1] "B represents a forest"
#> Log is not open.
#> [1] "B has a forest like structure"
#> Log is not open.
#> [1] "Root Root 0"
#> Log is not open.
#> [1] "Root Clone_1 1"
#> Log is not open.
#> [1] "Root Clone_2 0"
#> Log is not open.
#> [1] "Root Clone_3 0"
#> Log is not open.
#> [1] "Root Clone_4 0"
#> Log is not open.
#> [1] "Root Clone_5 0"
#> Log is not open.
#> [1] "Root Clone_6 0"
#> Log is not open.
#> [1] "Clone_1 Clone_1 1"
#> Log is not open.
#> [1] "Clone_1 Clone_2 1"
#> Log is not open.
#> [1] "Clone_1 Clone_3 0"
#> Log is not open.
#> [1] "Clone_1 Clone_4 0"
#> Log is not open.
#> [1] "Clone_1 Clone_5 0"
#> Log is not open.
#> [1] "Clone_1 Clone_6 0"
#> Log is not open.
#> [1] "Clone_2 Clone_2 1"
#> Log is not open.
#> [1] "Clone_2 Clone_3 1"
#> Log is not open.
#> [1] "Clone_2 Clone_4 1"
#> Log is not open.
#> [1] "Clone_2 Clone_5 0"
#> Log is not open.
#> [1] "Clone_2 Clone_6 0"
#> Log is not open.
#> [1] "Clone_3 Clone_3 1"
#> Log is not open.
#> [1] "Clone_3 Clone_4 1"
#> Log is not open.
#> [1] "Clone_3 Clone_5 1"
#> Log is not open.
#> [1] "Clone_3 Clone_6 0"
#> Log is not open.
#> [1] "Clone_4 Clone_4 1"
#> Log is not open.
#> [1] "Clone_4 Clone_5 1"
#> Log is not open.
#> [1] "Clone_4 Clone_6 0"
#> Log is not open.
#> [1] "Clone_5 Clone_5 1"
#> Log is not open.
#> [1] "Clone_5 Clone_6 1"
#> Log is not open.
#> [1] "Clone_6 Clone_6 1"
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#>      [,1]
#> [1,]    0
#> [2,]    1
#> [3,]    1
#> [4,]    1
#> [5,]    1
#> [6,]    1
#> [7,]    1
#> Log is not open.
#> [1] TRUE
#> Log is not open.
#> [1] "B represents a forest"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "Root is parent of Clone_1"
#> Log is not open.
#> [1] "Root is parent of Clone_2"
#> Log is not open.
#> [1] "Root is parent of Clone_3"
#> Log is not open.
#> [1] "Root is parent of Clone_4"
#> Log is not open.
#> [1] "Root is parent of Clone_5"
#> Log is not open.
#> [1] "Root is parent of Clone_6"
#> Log is not open.
#> [1] "B is a binary matrix"
#> Log is not open.
#> [1] "B is a square matrix"
#> Log is not open.
#> [1] "B is a full rank matrix"
#> Log is not open.
#> [1] "B is a forest "
#> Log is not open.
#> [1] "B has single root"
#> Log is not open.
#> [1] "all checks done!"
#> Log is not open.
#> [1] "continue..."
#> Log is not open.
#> [1] "clonal tree:"
#> Log is not open.
#>                        Root ARPC2_2_218249894_C_T PRAME_22_22551005_T_A
#> Root                      0                     1                     0
#> ARPC2_2_218249894_C_T     0                     0                     1
#> PRAME_22_22551005_T_A     0                     0                     0
#> COL1A2_7_94422978_C_A     0                     0                     0
#> HNRNPC_14_21211843_C_T    0                     0                     0
#> CCT8_21_29063389_G_A      0                     0                     0
#> RPL5_1_92837514_C_G       0                     0                     0
#>                        COL1A2_7_94422978_C_A HNRNPC_14_21211843_C_T
#> Root                                       0                      0
#> ARPC2_2_218249894_C_T                      0                      0
#> PRAME_22_22551005_T_A                      1                      1
#> COL1A2_7_94422978_C_A                      0                      0
#> HNRNPC_14_21211843_C_T                     0                      0
#> CCT8_21_29063389_G_A                       0                      0
#> RPL5_1_92837514_C_G                        0                      0
#>                        CCT8_21_29063389_G_A RPL5_1_92837514_C_G
#> Root                                      0                   0
#> ARPC2_2_218249894_C_T                     0                   0
#> PRAME_22_22551005_T_A                     0                   0
#> COL1A2_7_94422978_C_A                     1                   0
#> HNRNPC_14_21211843_C_T                    0                   0
#> CCT8_21_29063389_G_A                      0                   1
#> RPL5_1_92837514_C_G                       0                   0
#> Log is not open.
#>         Root ARPC2_2_218249894_C_T PRAME_22_22551005_T_A COL1A2_7_94422978_C_A
#> Root       1                     0                     0                     0
#> Clone_1    1                     1                     0                     0
#> Clone_2    1                     1                     1                     0
#> Clone_3    1                     1                     1                     1
#> Clone_4    1                     1                     1                     0
#> Clone_5    1                     1                     1                     1
#> Clone_6    1                     1                     1                     1
#>         HNRNPC_14_21211843_C_T CCT8_21_29063389_G_A RPL5_1_92837514_C_G
#> Root                         0                    0                   0
#> Clone_1                      0                    0                   0
#> Clone_2                      0                    0                   0
#> Clone_3                      0                    0                   0
#> Clone_4                      1                    0                   0
#> Clone_5                      0                    1                   0
#> Clone_6                      0                    1                   1
#> Log is not open.
#> [1] "No incongruences found due to the chronological order"
#> Log is not open.
#> [1] "first time mutation occurrences based the chronological order of samples:"
#> Log is not open.
#>    Root Clone_1 Clone_2 Clone_3 Clone_4 Clone_5 Clone_6 
#>       1       1       1       1       1       1       1 
#> Log is not open.
#> [1] "adjacent matrix with first time occurrences:"
#> Log is not open.
#>            Root_t1 Clone_1_t1 Clone_2_t1 Clone_3_t1 Clone_4_t1 Clone_5_t1
#> Root_t1          0          1          0          0          0          0
#> Clone_1_t1       0          0          1          0          0          0
#> Clone_2_t1       0          0          0          1          1          0
#> Clone_3_t1       0          0          0          0          0          1
#> Clone_4_t1       0          0          0          0          0          0
#> Clone_5_t1       0          0          0          0          0          0
#> Clone_6_t1       0          0          0          0          0          0
#>            Clone_6_t1
#> Root_t1             0
#> Clone_1_t1          0
#> Clone_2_t1          0
#> Clone_3_t1          0
#> Clone_4_t1          0
#> Clone_5_t1          1
#> Clone_6_t1          0
#> Log is not open.
#> [1] "time=" "1"    
#> Log is not open.
#> [1] "1st Root_t1->Root_t1=1, prev_Root_t1=0.143835616438356: Root_t1->Root_t2=2"
#> Log is not open.
#> [1] "1st Root_t1->Root_t2=2, prev_Root_t2=0.162162162162162: Root_t2->Root_t3=2"
#> Log is not open.
#> [1] "1st Root_t1->Clone_1_t1=1, prev_Clone_1_t1=0.157534246575342: Clone_1_t1->Clone_1_t2=2"
#> Log is not open.
#> [1] "1st Clone_1_t1->Clone_1_t2=2, prev_Clone_1_t2=0.225225225225225: Clone_1_t2->Clone_1_t3=2"
#> Log is not open.
#> [1] "1st Clone_1_t1->Clone_2_t1=1, prev_Clone_2_t1=0.184931506849315: Clone_2_t1->Clone_2_t2=2"
#> Log is not open.
#> [1] "1st Clone_2_t1->Clone_2_t2=2, prev_Clone_2_t2=0.243243243243243: Clone_2_t2->Clone_2_t3=2"
#> Log is not open.
#> [1] "1st Clone_2_t1->Clone_3_t1=1, prev_Clone_3_t1=0.280821917808219: Clone_3_t1->Clone_3_t2=2"
#> Log is not open.
#> [1] "1st Clone_2_t1->Clone_4_t1=1, prev_Clone_4_t1=0.0479452054794521: Clone_4_t1->Clone_4_t2=2"
#> Log is not open.
#> [1] "1st Clone_3_t1->Clone_3_t2=2, prev_Clone_3_t2=0.162162162162162: Clone_3_t2->Clone_3_t3=2"
#> Log is not open.
#> [1] "1st Clone_3_t1->Clone_5_t1=1, prev_Clone_5_t1=0.0958904109589041: Clone_5_t1->Clone_5_t2=2"
#> Log is not open.
#> [1] "1st Clone_4_t1->Clone_4_t2=2, prev_Clone_4_t2=0.0810810810810811: Clone_4_t2->Clone_4_t3=2"
#> Log is not open.
#> [1] "1st Clone_5_t1->Clone_5_t2=2, prev_Clone_5_t2=0.126126126126126: Clone_5_t2->Clone_5_t3=2"
#> Log is not open.
#> [1] "1st Clone_5_t1->Clone_6_t1=1, prev_Clone_6_t1=0.089041095890411: Clone_6_t1->Clone_6_t2=2"
#> Log is not open.
#> [1] "Clone_6_t1->Clone_6_t2=2, prev_Clone_6_t2=0: Clone_6_t1->Clone_6_t2=3"
#> Log is not open.
#> [1] "Clone_6_t1->Clone_6_t2=3, prev_Clone_6_t2=0: Clone_6_t2->Clone_6_t3=3"
#> Log is not open.
#> [1] "time=" "2"    
#> Log is not open.
#> [1] "1st Root_t2->Root_t3=2, prev_Root_t3=0.133333333333333: Root_t3->Root_t4=2"
#> Log is not open.
#> [1] "1st Clone_1_t2->Clone_1_t3=2, prev_Clone_1_t3=0.166666666666667: Clone_1_t3->Clone_1_t4=2"
#> Log is not open.
#> [1] "1st Clone_2_t2->Clone_2_t3=2, prev_Clone_2_t3=0.166666666666667: Clone_2_t3->Clone_2_t4=2"
#> Log is not open.
#> [1] "1st Clone_3_t2->Clone_3_t3=2, prev_Clone_3_t3=0.255555555555556: Clone_3_t3->Clone_3_t4=2"
#> Log is not open.
#> [1] "1st Clone_4_t2->Clone_4_t3=2, prev_Clone_4_t3=0.0333333333333333: Clone_4_t3->Clone_4_t4=2"
#> Log is not open.
#> [1] "1st Clone_5_t2->Clone_5_t3=2, prev_Clone_5_t3=0.211111111111111: Clone_5_t3->Clone_5_t4=2"
#> Log is not open.
#> [1] "1st Clone_6_t2->Clone_6_t3=3, prev_Clone_6_t3=0.0333333333333333: Clone_6_t3->Clone_6_t4=2"
#> Log is not open.
#> [1] "time=" "3"    
#> Log is not open.
#> [1] "time=" "4"    
#> Log is not open.
#> [1] "time=" "3"    
#> Log is not open.
#> [1] "time=" "2"    
#> Log is not open.
#> [1] "time=" "1"    
#> Log is not open.
#> [1] "idx"     "Clone_1"
#> Log is not open.
#> [1] "idx"     "Clone_2"
#> Log is not open.
#> [1] "idx"     "Clone_3"
#> Log is not open.
#> [1] "idx"     "Clone_4"
#> Log is not open.
#> [1] "idx"     "Clone_5"
#> Log is not open.
#> [1] "idx"     "Clone_6"
#> Log is not open.
#> [1] "/tmp/RtmpLExdXL/file5df03d88cab1/reference"
#> Log is not open.
#> [1] "info"
#> Log is not open.
#> [1] ""
#> Log is not open.

  

    

    

      

        

          
Longitudinal clonal tree

        

        

      

      

        

          
Fishplot